backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
redis==6.4.0
python-dotenv==1.1.1
httpx==0.28.1
SQLAlchemy==2.0.44
psycopg2-binary==2.9.10
